Modify bytecode interpreter instruction processing

+++ mathparser.cpp
@@ -3,5 +3,5 @@
 Char op[NUMOP]={'+','-','*','/','^'};
 String funcName[NUMFUNC]={"sin","cos","tan","exp","pow"};
 int funcNumArgs[NUMFUNC]={1,1,1,1,2};
-long opcode[NUMOP]={ADD,SUB,MULT,DIV,POW_N};
-long funcCode[NUMFUNC]={SIN,COS,TAN,EXP,POW};
+int opcode[NUMOP]={ADD,SUB,MULT,DIV,POW_N};
+int funcCode[NUMFUNC]={SIN,COS,TAN,EXP,POW};
